Extract `Format` to file
